Added clang tools and CMake changes

+++ include/alrightData.h
@@ -11,9 +11,8 @@ namespace alright {
 
 class headerField {
  public:
-  explicit headerField(std::string aName, std::string aValue)
-      : name(aName)
-      , value(aValue) {}
+  explicit headerField(std::string aName, std::string aValue) : name(aName), value(aValue) {}
+
  private:
   std::string name;
   std::string value;
@@ -21,32 +20,28 @@ class headerField {
 
 class serverEndpointData {
  public:
-  explicit serverEndpointData(unsigned int aPortNumber, std::string aServerPath)
-      : portNumber(aPortNumber)
-      , serverPath(aServerPath) {}
-  const unsigned int portNumber;
+  explicit serverEndpointData(short unsigned int aPortNumber, std::string aServerPath)
+      : portNumber(aPortNumber), serverPath(aServerPath) {}
+  const short unsigned int portNumber;
   const std::string serverPath;
 };
 
 class clientEndpointData {
  public:
-  explicit clientEndpointData(unsigned int aPN,
-                              std::string aAddress,
-                              std::string aRequestMethod,
-                              std::string aRequestURI,
-                              std::string aRequestHttpVersion)
-  : portNumber(aPN)
-  , address(aAddress)
-  , requestMethod(aRequestMethod)
-  , requestURI(aRequestURI) 
-  , requestHttpVersion(aRequestHttpVersion) {}
-  const unsigned int portNumber;
+  explicit clientEndpointData(short unsigned int aPN, std::string aAddress, std::string aRequestMethod,
+                              std::string aRequestURI, std::string aRequestHttpVersion)
+      : portNumber(aPN),
+        address(aAddress),
+        requestMethod(aRequestMethod),
+        requestURI(aRequestURI),
+        requestHttpVersion(aRequestHttpVersion) {}
+  const short unsigned int portNumber;
   const std::string address;
   const std::string requestMethod;
   const std::string requestURI;
   const std::string requestHttpVersion;
 };
 
-} // namespace alright
+}  // namespace alright
 
 #endif
